% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestedLoops.R
\name{nestedLoopsRcpp}
\alias{nestedLoopsRcpp}
\title{Title}
\usage{
nestedLoopsRcpp(x, y)
}
\arguments{
\item{x}{description}

\item{y}{description}
}
\value{
description
}
\description{
Title
}
\examples{
1
}
